remove test comments and dont cover operation utils
